+++ v1/pkg/client/model_host_all_of.go
@@ -55,7 +55,7 @@ type HostAllOf struct {
 	Substate HostSubstate `json:"Substate,omitempty"`
 	StateTime time.Time `json:"StateTime,omitempty"`
 	SubstateTime time.Time `json:"SubstateTime,omitempty"`
-	Progress int32 `json:"Progress,omitempty"`
+	Progress uint64 `json:"Progress,omitempty"`
 	Alert bool `json:"Alert,omitempty"`
 	AlertInfo []HostAlertInfo `json:"AlertInfo,omitempty"`
 }
