+++ v1/pkg/client/model_host_usage_entry.go
@@ -31,7 +31,7 @@ type HostUsageEntry struct {
 	// The end of the usage reporting window or when the resource was freed
 	UsageEnd time.Time `json:"UsageEnd,omitempty"`
 	// The difference between the UsageEnd and UsageStart rounded up to the UsageHours
-	UsageHours int32 `json:"UsageHours,omitempty"`
+	UsageHours uint64 `json:"UsageHours,omitempty"`
 	// Description of error that affected the usage reporting
 	Error string `json:"Error,omitempty"`
 	// Name of the MachineSize requested when host was created
