+++ v1/pkg/client/model_volume_usage_entry_all_of.go
@@ -22,5 +22,5 @@ type VolumeUsageEntryAllOf struct {
 	// Unique ID of the volume flavor used when creating the volume
 	FlavorID string `json:"FlavorID,omitempty"`
 	// The size of the volume in MB
-	Capacity int32 `json:"Capacity,omitempty"`
+	Capacity uint64 `json:"Capacity,omitempty"`
 }
